feat(scores): extract pure-logic helpers to ScoreResolverService
New @Service mirroring SequenceValidationService — no own state, autowires
PilotService. Houses the resolver's pure-logic helpers so the controller
stays a thin HTTP layer.

Moved verbatim from APIController:
- getMismatches()             — body of /api/scores/mismatches
- countRoundsForType(...)     — utility used by every resolver eligibility
                                check; callers in moveRound() redirect to
                                the service
- findUnresolvedMissingSeq2Round(Pilot, List<Pilot>)
                              — per-pilot peer-evidence check
- checkAndAddMismatches(...)  — private detection helper

The /api/scores/mismatches endpoint becomes thin delegation. APIController
sheds the now-unused Comparator / HashSet / Set / ContestClasses imports.
